Fix DataFormProvider parsing 'fixed' FormFields
b71039660bf73fb537bfafe09f39c993ba84f1c1 made FormField.setType(Type)
throw an illegal argument exception if type is fixed, but the
DataFormProvider was not changed, so it still would call
setType(Type.fixed).

Change DataFormProvider so that the non-argument constructer of
FormField is used when type == fixed.